Add barebones non-null-terminated string library
Non-null-terminated strings are incredibly useful for a vfs becuase they
can be substringed without calling memcpy(...) or inserting a null
terminator (at the end of the substring).  The plan is to use
kstr_strtok(...) in filepath resolution and kstr_hash(...) in the
hashmaps that make up direntry children lists.

Misc:
* The updated QRUN was stolen from my console attempt
(dev/aw8465/console; commit 8f08683)