Add slab cache
These caches store and manage arbitrarily sized, homogenously sized
chunks of memory. They're useful for making in memory caches of
important structures. This is exactly what they're intended to be used
for (as of right now): the inode and dnode caches (as well as maybe
filename strings, but that would likely be better served with a
heterogenous allocator such as a buddy allocator).

They are built on top of the existing os's page and slice allocators and
can use either sized chunk of memory as its slabs. Smaller slabs save
memory when there are fewer elements, whereas larger slabs are generally
faster for larger numbers of elements (they allocate new slabs [call
_km_page_alloc()] less frequently).