Implement UNTESTED fopen syscall
This required adding open file tables to pcbs. They are stored in a slab
cache and set to a fixed size (configurable in params.h). The function
to get a pcb's next available file descriptor _pcb_get_next_fd(...) is
also located in procs.c to maintain locality.

When allocating a new pcb, an open file table is grabbed from the cache
(cleared to 0) and assigned to pcb->open_files. stdin and stdout
inheritance (fds 0 and 1) has not yet been implemented yet (in the fork
syscall) but a TODO has been added to the future location.

The open implementation also necessitated the creation of a kfile_t
cache (to hold all open files). This is also a slab cache, and stored in
vfs.c (although the public interface is _vfs_allocate_file(void).