Fix cio read syscall bug and revert testing code
Bug description copied from email to Carithers:

---

In the read syscall, the cio ring buffer is first checked to see if
there is any data, if there is none read() fails fast and returns
e_no_data. However, if there is data, it proceeds into __cio_reads().
This is where the problem lies. __cio_reads() assumes that the buffer
it is recieving is of minimum length 2 (without checking properly) and
at cio.c:661, it will skip reading from the ring buffer if the buffer
length is 1.

This leaves the read syscall in an unexpected state: the cio ring buffer
is not empty, but __cio_reads() still returned n=0 bytes read. At this
stage, the read syscall assumes that n=0 is only possible if the serial
channel is selected and no input was available so it puts the current
process on the sio queue (instead of just returning 0 or e_no_data).

---

The fix I implemented is to merely check whether the buffer passed to
a cio channel read syscall is < 2 (and if so return an error code). This
doesn't fix the underlying issues, but it prevents the bug from cropping
up in cio reads from userland and preventing processes to inexplicably
hang / get swapped out ("WE DO NOT BREAK USERSPACE!").

I also removed a bunch of debugging code from both the clock isr (it
no longer calls _pcb_dump) and the shell itself.

I also also moved the wtsh.c file to the updated source tree structure
(src/userspace/wtsh.c -> src/usr/userspace/wtsh.c).