Add userland prog for vfs testing and lots more
Honestly adding the userland program was the least of the changes made
here. The main goal was to test the fopen syscall, but for that to work,
a lot of other changes had to be made. This basically amounts to the
beginning of the public (kernel not userland) interface to the module.
However, at the end of the day, I was able to sucessfully test and run
fopen on multiple different files in a fs driver written to the vfs
interface.

Below is a (probably not exhaustive) list of all the changes made:

* Added a function to register filesystem types (and a store in vfs.c)

* There is a function to mount a filesystem type at a specific path
(although the path is currently ignored). The fs_type_num parameter
corresponds to the fs type the caller wants to mount and is assigned
incrementally as filesystems are registered (this isn't a great way to
do it and will hopefully change in the future)

* Added slab caches to vfs.c for dirents, superblocks, and mounts (with
associated public allocate functions for each type)

* Moved all but the backing data structures in testfs to the new dynamic
memory system (it's slab caches all the way down)