Implement fclose syscall
This also adds a close kf_op which is an callback into the inode /
file's driver (for performing actions like freeing the private context,
flushing to disk, etc.) however it is _COMPLETELY OPTIONAL_. If the
function isn't supplied, flclose will still clear the entry in
_current's open file table and free the file struct. It would be kind of
funny to have a driver that supports opening files but forces them to
stay open though. Food for future kernel developers' thoughts.