Add fopen modes, read/write locks, and inode types
This is an incremental commit in preperation for fread and fwrite.
First, open files now have modes: O_READ and O_WRITE (or O_RDWR). Their
semantics are identical to those in POSIX. In the future, only files
opened with the appropriate mode will be allowed to read or write. A fun
little quirk I left in intentionally is that directories are allowed to
be opened for writing. It will be up to the fs driver's writer what the
semantics of writing to a directory are but it seems like a fun thing to
discover.

Next, for the file locking, there are two types of locks: read locks and
write locks. Many readers can have the same inode open, however write
locks are exclusive. When a process opens an inode for writing, no other
readers _OR_ writers can open the file. Likewise, if there are any
readers currently reading a file, a writer cannot open it. All locks are
mandatory (automatically enforced) and device inodes are never locked
(to allow for sharing of a tty device later). I.e. any number of readers
and writers can have a device inode open at the same time: race
conditions be damnned.

Finally there are now three types of inode (stored in a type variable:
_NOT_ new structures / types) S_TYPE_FILE, S_TYPE_DIR, S_TYPE_DEV. Their
meanings are self explanitory. FS drivers are responsible for assigning
the appropriate types to each inode when creating or loading them (from
disk).