Implement write syscall
For testing, I had to alter the way testfs gets its data backing
(Directories are excluded from the following). For every file (leaf node
in the bogus tree), a page of memory is allocated. This is the backing
for file data (and files cannot be grown past it). However, each node
also keeps track of its file "length" to emulate a real filesystem that
can grow files beyond their current size by writing past the end. For
testfs, any writes past the size of a page (4096) will fail with error
E_NO_MEM.