Refactor namey.c and properly build direntry cache
There are several cases where the direntry for a path is needed (instead
of the inode it refers to) so the path decoding logic was refactored out
of namey(...) to resolve_path(...). namey(...) is essentially a wrapper
(with a different interface and some basic error checking) for
resolve_path(...).

Additionally, the direntry cache is now properly built and functional.
When resolving a path, resolve_path(...) first checks the children of
the current direntry for the current path component. If there is a
positive entry, it becomes the current direntry and the next path
component is resolved. If there is a negative entry, the search fails
then and there (avoiding a "costly" call to lookup. Finally, if there is
no entry, a new direntry_t is allocated (with the current path component
as its name), lookup is called on the current direntry's inode (with the
new direntry as a parameter), and the filled (or negative) direntry is
inserted into the direntry (tree) hirearchy.

There will never be a reason to deallocate / delete a direntry (and
therefore reorder or manipulate the tree structure), because even if an
inode is deleted, all direntries referring to it become negative (not
deleted).