Update to build and run on modern machines
The original code has a global variable defined in `vgatext.h` which
caused the variable to be defined in multiple code units. Whichever
version of `ld` was installed on the school lab machines didn't complain
but `ld 2.40` fails with an error.

Additionally, the qemu option `-curses` has been removed between the lab
machines' version and qemu 10.0.0 meaning `-curses` needed to be updated
to `-display curses`.